feat✨(system): implement comprehensive machine monitoring and client management

+++ src/main/java/cv/beriholic/beeyes/models/dto/system/CPUInfo.java
@@ -32,5 +32,5 @@ public class CPUInfo implements Serializable {
     /**
      * CPU使用率
      */
-    private Float usage;
+    private Double usage;
 }
